+++ seed/static/seed/js/controllers/update_derived_data_modal_controller.js
@@ -15,9 +15,8 @@ angular.module('SEED.controller.update_derived_data_modal', []).controller('upda
     $scope.taxlot_view_ids = _.uniq(taxlot_view_ids);
 
     $scope.begin_update = () => {
-      inventory_service.update_derived_data($scope.property_view_ids).then($uibModalInstance.close);
+      inventory_service.update_derived_data($scope.property_view_ids, $scope.taxlot_view_ids).then($uibModalInstance.close);
     };
-
 
 
     /**
